# Resolve #17 optimize use of '^' / '\A'