The problem was that a regular expression like `^x` would take linear time to is_match a string `aaa....` when all it needed to do was check the first character then bail out. This PR makes it so if the regex begins with `^` (without multiline) or `\A`, then it stops iterating over characters as soon as there are no possible matches.

Benchmarks:

```
dynamic:
anchored_literal_long_non_match: 7492 ns/iter => 435 ns/iter
anchored_literal_short_non_match: 917 ns/iter => 431 ns/iter

native:
anchored_literal_long_non_match: 5020 ns/iter => 50 ns/iter
anchored_literal_short_non_match: 375 ns/iter => 50 ns/iter
```