@huonw 2015-01-14
Thanks!

Could I ask you to expand the commit message a little to describe what the problem is and how it is fixed? Just saying the issue number is better than nothing, but it does mean that someone looking the source and checking the git history needs to cross reference against github (which may have disappeared at that point).

(I'm happy to help if you need assistance with git or anything else.)