@tmerr 2015-01-14
Sure, expanded. Is that alright?