@BurntSushi 2015-01-14
@tmerr This looks great. I'm happy to see that the change is small. I just had one question, but otherwise this looks awesome. Thank you!