@huonw 2015-01-14
@tmerr that messsage looks much better, thanks! Could you place it into the git commit message (e.g. `git commit --amend` and then `git push -f` to update the PR by force pushing over this branch.)