@tmerr 2015-01-14
@huonw Done, thanks for the help!

It's been a blast reading through this code by the way. I had been aware of NFAs but only  the "draw pretty circles on paper with lines between them" version, not as a machine with split and jump instructions. The two models of thinking about it meld together nicely