@huonw 2015-01-14
Thanks! I merged as 10edff6bb76f1c9f4423f3bb8ad373d0c7a69773 (there was a piece of trailing whitespace which I took the initiative to remove).